custom-foreign-trade/requirements.txt
flask
gunicorn
requests
Werkzeug
PyJWT==1.7.1
Jinja2==2.11.2
MarkupSafe==1.1.1
idna==2.10
itsdangerous==1.1.0
click==7.1.2
chardet==3.0.4
urllib3==1.26.2